% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plans_oc.R
\name{compare_plans_oc}
\alias{compare_plans_oc}
\title{Comparison based on OC curve}
\usage{
compare_plans_oc(c1, c2, c3, r1, t1, r2, t2, r3, t3, distribution, K, m, sd)
}
\arguments{
\item{c1, c2, c3}{acceptance numbers}

\item{r1, r2, r3}{number of primary increments in a grab sample or grab sample size}

\item{t1, t2, t3}{number of grab samples}

\item{distribution}{what distribution we have used such as   \code{'Poisson gamma'} or \code{'Lognormal'}or \code{'Poisson lognormal'}}

\item{K}{dispersion parameter of the Poisson gamma distribution (default value 0.25)}

\item{m}{microbiological limit with default value zero, generally expressed as number of microorganisms in specific sample weight}

\item{sd}{standard deviation of the lognormal and Poisson-lognormal distributions on the log10 scale (default value 0.8)}
}
\description{
This function produces overlaid Operating Characteristic (OC) curves for any three systematic/random sampling schemes for specified parameters.
}
\examples{
c1 <- 0
c2 <- 0
c3 <- 0
r1 <- 25
r2 <- 50
r3 <- 75
t1 <- 10
t2 <- 10
t3 <- 10
distribution <- 'Poisson lognormal'
\donttest{compare_plans_oc(c1, c2, c3, r1, t1, r2, t2, r3, t3, distribution)}
\donttest{compare_plans_oc(c1, c2, c3, r1, t1, r2, t2, r3, t3, distribution, sd = 0.4)}
}
\references{
\itemize{
\item Jongenburger, I., Besten, H.M., & Zwietering, M.H. (2015). Statistical aspects of food safety sampling. Annual review of food science and technology, 6, \href{https://doi.org/10.1146/annurev-food-022814-015546}{479-503}.
\item Mussida, A., Vose, D. & Butler, F. Efficiency of the sampling plan for {C}ronobacter spp. assuming a Poisson lognormal distribution of the bacteria in powder infant formula and the implications of assuming a fixed within and between-lot variability, Food Control, Elsevier, 2013 , 33 , \href{https://doi.org/10.1016/j.foodcont.2013.02.021}{174-185}.
\item Van Schothorst, M., Zwietering, M., Ross, T., Buchanan, R. & Cole, M., Relating microbiological criteria to food safety objectives and performance objectives  Food Control , 2009 , 20 , \href{https://doi.org/10.1016/j.foodcont.2008.11.005}{967-979}.
}
}
\seealso{
\link{prob_accept}
}
