% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plans_oc_sd.R
\name{compare_plans_oc_sd}
\alias{compare_plans_oc_sd}
\title{Comparison based on OC curve with different standard deviations}
\usage{
compare_plans_oc_sd(c1, c2, c3, r1, t1, r2, t2, r3, t3, sd1, sd2, sd3, distribution, K, m)
}
\arguments{
\item{c1, c2, c3}{acceptance numbers}

\item{r1, r2, r3}{number of primary increments in a grab sample or grab sample size}

\item{t1, t2, t3}{number of grab samples}

\item{sd1, sd2, sd3}{standard deviations of the lognormal and Poisson-lognormal distributions on the log10 scale}

\item{distribution}{what distribution we have used such as   \code{'Poisson gamma'} or \code{'Lognormal'}or \code{'Poisson lognormal'}}

\item{K}{dispersion parameter of the Poisson gamma distribution (default value 0.25)}

\item{m}{microbiological limit with default value zero, generally expressed as number of microorganisms in specific sample weight}
}
\value{
overlaid OC curves
}
\description{
This function produces overlaid Operating Characteristic (OC) curves for any three systematic/random sampling schemes for specified parameters with different standard deviation vlues.
}
\examples{
c1 <- 0
c2 <- 0
c3 <- 0
r1 <- 25
r2 <- 25
r3 <- 25
t1 <- 30
t2 <- 30
t3 <- 30
sd1 <- 0.2
sd2 <- 0.4
sd3 <- 0.8
distribution <- 'Poisson lognormal'
compare_plans_oc_sd(c1, c2, c3, r1, t1, r2, t2, r3, t3, sd1, sd2, sd3, distribution)
}
\seealso{
\link{prob_accept}
}
